samples/adobe-jwt-python/requirements.txt
PyJWT~=2.6.0
requests~=2.28.1
cryptography~=38.0.4
